Expand InterfaceData API. Allow better signal/method inspection.
Add InterfaceData methods

    NumMethod() int
    Method(i int) MethodData
    MethodByName(name string) MethodData

Also add corresponding Signal methods.

This API is more consistent with the APIs of existing packages in the Go
standard library.

Although, due to a 'feature' of the xml package, the internal
InterfaceData types have gotten somewhat more complex. There is now an
xmlInterfaceData object (which is marshled/unmarshaled from/to), along
with type interfaceData, which is actually handled when the
GetInterfaceData method is called. There is a note about this as a
godoc comment for type xmlInterfaceData.